requirements.txt
python-dotenv
requests
fastapi
uvicorn
pydantic[email]
gspread
oauth2client
google-auth
resend